+++ galacteek/core/glogger.py
@@ -1,17 +1,34 @@
+import os.path
+import asyncio
 import sys
 import collections
+import platform
+import traceback
+from datetime import datetime
+
 from cachetools import cached
 from cachetools import LRUCache
 from colour import Color
 
 from logbook import Logger, StreamHandler
+from logbook import Handler
 from logbook import TimedRotatingFileHandler
+from logbook.handlers import StringFormatterHandlerMixin
 from logbook.more import ColorizedStderrHandler
 from logbook.helpers import u
 from logbook.compat import redirect_warnings
 from logbook.compat import redirect_logging
+from logbook.base import _datetime_factory
 
 from galacteek.core import SingletonDecorator
+
+try:
+    from aiofile import AIOFile
+except ImportError:
+    haveAioFile = False
+else:
+    haveAioFile = True
+
 
 mainFormatString = u(
     '[{record.time:%Y-%m-%d %H:%M:%S.%f%z}] '
@@ -22,10 +39,14 @@ easyFormatString = u(
     '{record.module}: {record.message}')
 
 
-@SingletonDecorator
+loggerMain = Logger('galacteek')
+loggerUser = Logger('galacteek.user')
+
+
+@ SingletonDecorator
 class LogRecordStyler:
     red = Color('red')
-    darkred = Color('red')
+    darkred = Color('darkred')
     blue = Color('blue')
     turquoise = Color('turquoise')
     green = Color('green')
@@ -148,13 +169,16 @@ class LogRecordStyler:
         }
     }
 
-    @cached(LRUCache(64))
+    @ cached(LRUCache(64))
     def color(self, color, r=0, g=0, b=0):
         try:
             c = Color(color)
             c.red += r
             c.green += g
             c.blue += b
+        except ValueError as verr:
+            loggerMain.warning(f'{self.__class__}: Cannot build color: {verr}')
+            return self._defaultColor
         except Exception:
             return self._defaultColor
         else:
@@ -165,18 +189,18 @@ class LogRecordStyler:
 
         def _g(st):
             cName = st.get('basecolor')
-            r, g, b = st.get('red', 0), st.get('green', 0), \
-                st.get('blue', 0)
+            r, g, b = st.get('red', 0), st.get('green', 0), st.get('blue', 0)
 
             return self.color(cName, r, g, b), None
 
         if style:
             return _g(style)
         else:
-            parent = '.'.join(record.module.split('.')[:-1])
-            style = self._table.get(parent)
-            if style:
-                return _g(style)
+            if record.module:
+                parent = '.'.join(record.module.split('.')[:-1])
+                style = self._table.get(parent)
+                if style:
+                    return _g(style)
 
             # default
             return self._defaultColor, None
@@ -215,11 +239,114 @@ class ColorizedHandler(ColorizedStderrHandler):
             return 'lightgray'
 
 
+class AsyncLogHandler(Handler, StringFormatterHandlerMixin):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        Handler.__init__(self)
+        StringFormatterHandlerMixin.__init__(self, None)
+
+        self.loop = asyncio.get_event_loop()
+        self.ptask = self.loop.create_task(self._process())
+        self.__queue = asyncio.Queue()
+
+    def emit(self, record):
+        if record.module:
+            self.__queue.put_nowait(record)
+
+    async def _process(self):
+        while True:
+            try:
+                record = await self.__queue.get()
+                await self.processRecord(record)
+                self.__queue.task_done()
+            except asyncio.CancelledError:
+                return
+            except Exception:
+                return
+
+    async def processRecord(self, record):
+        pass
+
+
+class AsyncDatedFileHandler(AsyncLogHandler):
+    def __init__(self, filename,
+                 mode='a', encoding='utf-8', level='debug',
+                 format_string=None, date_format='%Y-%m-%d',
+                 backup_count=0, filter=None, bubble=False,
+                 timed_filename_for_current=True,
+                 rollover_format='{basename}-{timestamp}{ext}'):
+        self.format_string = format_string
+        AsyncLogHandler.__init__(self)
+
+        self.date_format = date_format
+        self.backup_count = backup_count
+        self.format_string = format_string
+        self.encoding = encoding
+
+        self.rollover_format = rollover_format
+
+        self.original_filename = filename
+        self.basename, self.ext = os.path.splitext(os.path.abspath(filename))
+        self.timed_filename_for_current = timed_filename_for_current
+
+        self._timestamp = self._get_timestamp(_datetime_factory())
+        if self.timed_filename_for_current:
+            filename = self.generate_timed_filename(self._timestamp)
+        elif os.path.exists(filename):
+            self._timestamp = self._get_timestamp(
+                datetime.fromtimestamp(
+                    os.stat(filename).st_mtime
+                )
+            )
+
+        self._aio_fd = AIOFile(filename, 'w+t')
+        self._aio_position = 0
+
+    @property
+    def afd(self):
+        return self._aio_fd
+
+    def _get_timestamp(self, datetime):
+        return datetime.strftime(self.date_format)
+
+    def generate_timed_filename(self, timestamp):
+        """
+        Produces a filename that includes a timestamp in the format supplied
+        to the handler at init time.
+        """
+        timed_filename = self.rollover_format.format(
+            basename=self.basename,
+            timestamp=timestamp,
+            ext=self.ext)
+        return timed_filename
+
+    async def processRecord(self, record):
+        try:
+            await self.afd.open()
+
+            formatted = self.format(record) + "\n"
+
+            await self.afd.write(
+                formatted,
+                offset=self._aio_position
+            )
+            await self.afd.fsync()
+
+            self._aio_position += len(formatted)
+        except Exception:
+            traceback.print_exc()
+
+
 def basicConfig(outputFile=None, level='INFO',
-                redirectLogging=False, colorized=False):
+                redirectLogging=False, colorized=False,
+                loop=None):
     if outputFile:
-        handler = TimedRotatingFileHandler(outputFile,
-                                           date_format='%Y-%m-%d')
+        if platform.system() == 'Linux':
+            handler = AsyncDatedFileHandler(outputFile,
+                                            date_format='%Y-%m-%d')
+        else:
+            handler = TimedRotatingFileHandler(outputFile,
+                                               date_format='%Y-%m-%d')
     else:
         if not colorized:
             handler = StreamHandler(sys.stderr, level=level, bubble=True)
@@ -233,7 +360,3 @@ def basicConfig(outputFile=None, level='INFO',
     if redirectLogging:
         redirect_logging()
         redirect_warnings()
-
-
-loggerMain = Logger('galacteek')
-loggerUser = Logger('galacteek.user')