+++ galacteek/ui/videocall.py
@@ -1,5 +1,6 @@
 import orjson
 
+from galacteek import log
 from galacteek.ipfs import ipfsOp
 from galacteek.ipfs.cidhelpers import IPFSPath
 
@@ -14,33 +15,37 @@ from .dialogs import VideoChatAckWait
 class VideoCallInitiator:
     def __init__(self, ipService):
         self.app = runningApp()
-        self.ipService  = ipService
+        self.ipService = ipService
 
     @ipfsOp
     async def start(self, ipfsop):
         await self.videoRendezVous(ipfsop)
 
     async def videoRendezVous(self, ipfsop):
-        remotePeerId, serviceName = ipfsop.p2pEndpointExplode(
-            self.ipService.endpoint)
+        # remotePeerId, serviceName = ipfsop.p2pEndpointExplode(
+        #     self.ipService.endpoint)
+
+        endpointAddr = await self.ipService.p2pEndpointAddr()
 
         req = {
             'peer': ipfsop.ctx.node.id
         }
 
+        log.debug(f'Req {req} to {endpointAddr}')
+
         try:
-            async with ipfsop.p2pDialer(
-                    remotePeerId, serviceName,
+            async with ipfsop.p2pDialerFromAddr(
+                    endpointAddr,
                     addressAuto=True) as sCtx:
                 if sCtx.failed:
                     raise Exception(f'Cannot reach {remotePeerId}')
 
                 async with sCtx.session.ws_connect(
-                    sCtx.httpUrl('/rendezVousWs')) as ws:
+                        sCtx.httpUrl('/rendezVousWs')) as ws:
 
                     await self.rendezVousMain(ipfsop, ws)
         except Exception as err:
-            print(str(err))
+            log.debug(str(err))
             return None
 
     async def rendezVousMain(self, ipfsop, ws):
@@ -52,7 +57,7 @@ class VideoCallInitiator:
         await ipfsop.sleep()
         ackWaitWidget = VideoChatAckWait()
 
-        #msg = await ws.receive_json()
+        # msg = await ws.receive_json()
         async for rawmsg in ws:
             try:
                 msg = orjson.loads(rawmsg.data)
@@ -63,7 +68,7 @@ class VideoCallInitiator:
             if msg['msgtype'] == MSGTYPE_CAPTCHA_CHALLENGE:
                 await ipfsop.sleep()
 
-                sessionId = msg.get(MSGF_SESSIONID)
+                # sessionId = msg.get(MSGF_SESSIONID)
                 captchaCid = msg.get(MSGF_CAPTCHACID)
 
                 captchaRaw = await ipfsop.catObject(captchaCid)
@@ -86,19 +91,20 @@ class VideoCallInitiator:
                         break
 
             elif msg['msgtype'] == MSGTYPE_ACKWAIT:
-                #await runDialogAsync(dlg)
+                # await runDialogAsync(dlg)
                 ackWaitWidget.show()
 
             elif msg['msgtype'] == MSGTYPE_RENDEZVOUS:
                 ackWaitWidget.hide()
                 rvTopic = msg.get(MSGF_RVTOPIC)
-                print(rvTopic)
+
+                log.debug(f'Rendez-vous topic {rvTopic}')
 
                 await self.startVideoCall(rvTopic)
 
     @ipfsOp
     async def startVideoCall(self, ipfsop, rvTopic):
-        rootPath = IPFSPath(ipfsop.ctx.resources['videocall']['Hash'])
+        rootPath = IPFSPath(ipfsop.ctx.resources['app-videocall']['Hash'])
         offerPath = rootPath.child('offer.html')
         offerPath.fragment = rvTopic